soft delete resources

+++ src/Action/Api/Resource/DeleteAction.php
@@ -78,7 +78,18 @@ final class DeleteAction implements MiddlewareInterface
         /** @var EntityInterface $entity */
         $entity = $repository->find($request->getAttribute("id"));
 
-        $repository->remove($entity);
+        if (method_exists($entity, 'deletedAt')) {
+            //if(!$entity->deletedAt()) {
+            $repository->save($entity->with('deletedAt', new \DateTimeImmutable()));
+            /**
+             * TODO: implement permanent deletion
+             */
+            //} else {
+            //$repository->remove($entity);
+            //}
+        } else {
+            $repository->remove($entity);
+        }
 
         return new ApiSuccessResponse();
     }
